// frontend/src/api/generated/models/dashboardDataDoc.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * acpms-server
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardAgentLogDoc } from './dashboardAgentLogDoc';
import type { DashboardHumanTaskDoc } from './dashboardHumanTaskDoc';
import type { DashboardProjectDoc } from './dashboardProjectDoc';
import type { DashboardStatsDoc } from './dashboardStatsDoc';

export interface DashboardDataDoc {
  agentLogs: DashboardAgentLogDoc[];
  humanTasks: DashboardHumanTaskDoc[];
  projects: DashboardProjectDoc[];
  stats: DashboardStatsDoc;
}